pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "boo-backend"
version = "0.1.0"
description = "BOO mascot growth backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.2",
  "pydantic-settings>=2.3.4",
  "passlib>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.2",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
addopts = "-q"

[tool.setuptools.packages.find]
include = ["app*"]
